# [Additions] Steam Achievement Progress Updating
- Steam achievements can now be tracked properly
- Added an extra variable for the steam stat to allow users to update steam achievements with them
- Users may now choose between ints or floats when uploading their achievement progress to other platforms (just Steam for now)
- Progress is now stored as a float so it can be used as either an int or float)

+++ Plugins/AchievementPlugin/Source/AchievementPlugin/Private/AchievementPluginBPLibrary.cpp
@@ -43,7 +43,7 @@ UAchievementManagerSubSystem* GetManager()
 //	return TArray<FString>();
 //}
 
-bool UAchievementPluginBPLibrary::IncreaseAchievementProgress(const FString& localAchievementId, const int32 increase)
+bool UAchievementPluginBPLibrary::IncreaseAchievementProgress(const FString& localAchievementId, const float increase)
 {
 	return GetManager()->IncreaseAchievementProgress(localAchievementId, increase);
 }
