+++ Plugins/AchievementPlugin/Source/AchievementPlugin/Public/AchievementPluginBPLibrary.h
@@ -25,7 +25,7 @@ public:
 	UFUNCTION(BlueprintCallable, meta = (DisplayName = "Set Achievement Progress", Keywords = "Set Achievement Progress"), Category = "AchievementPlugin")
 	static bool IncreaseAchievementProgress(
 		const FString& localAchievementId,
-		int32 increase);
+		float increase);
 
 	UFUNCTION(BlueprintCallable, meta = (DisplayName = "Save Achievement Progress Async", Keywords = "Save Achievement Progress Async"), Category = "AchievementPlugin")
 	static bool SaveAchievementProgressAsync();
